omap4: hwmod_data: add missing fields for fdif
The FDIF hwmod data for OMAP4 is missing the chip version and
the clock domain name fields. The absence of chip name caused
the hwmod to be ignored altogether during registration, and
the lack of the clock domain caused the hwmod setup to fail,
rendering fdif use-cases non-functional.

This missing data has been added.

Signed-off-by: Chandra Sekhar.Anagani <chandu@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>